requirements.txt
discord.py
PyNaCl
yt-dlp
python-dotenv
requests
ffmpeg
Flask
google-api-python-client
aiohttp
